move callback flow to ProductView
